+++ BambooTracker/gui/comment_edit_dialog.cpp
@@ -31,7 +31,7 @@ CommentEditDialog::CommentEditDialog(QString comment, QWidget *parent) :
 	ui(new Ui::CommentEditDialog)
 {
 	ui->setupUi(this);
-	setWindowFlags(windowFlags() ^ Qt::WindowContextHelpButtonHint);
+	setWindowFlags(windowFlags() & ~Qt::WindowContextHelpButtonHint);
 	ui->plainTextEdit->setPlainText(comment);
 }