+++ BambooTracker/gui/vgm_export_settings_dialog.cpp
@@ -34,7 +34,7 @@ VgmExportSettingsDialog::VgmExportSettingsDialog(QWidget *parent) :
 {
 	ui->setupUi(this);
 
-	setWindowFlags(windowFlags() ^ Qt::WindowContextHelpButtonHint);
+	setWindowFlags(windowFlags() & ~Qt::WindowContextHelpButtonHint);
 
 	for (QRadioButton *button : {
 		 ui->ym2608RadioButton, ui->ym2612RadioButton, ui->ym2203RadioButton, ui->noneFmRadioButton,