+++ BambooTracker/gui/wave_export_settings_dialog.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018-2019 Rerrah
+ * Copyright (C) 2018-2021 Rerrah
  *
  * Permission is hereby granted, free of charge, to any person
  * obtaining a copy of this software and associated documentation
@@ -25,18 +25,40 @@
 
 #include "wave_export_settings_dialog.hpp"
 #include "ui_wave_export_settings_dialog.h"
+#include <algorithm>
+#include <QListWidgetItem>
+#include "gui/gui_utils.hpp"
+#include "song.hpp"
 
-WaveExportSettingsDialog::WaveExportSettingsDialog(QWidget *parent) :
+WaveExportSettingsDialog::WaveExportSettingsDialog(const std::vector<int> defUnmutes, QWidget *parent) :
 	QDialog(parent),
 	ui(new Ui::WaveExportSettingsDialog)
 {
 	ui->setupUi(this);
 
-	setWindowFlags(windowFlags() ^ Qt::WindowContextHelpButtonHint);
+	setWindowFlags(windowFlags() & ~Qt::WindowContextHelpButtonHint);
 
 	ui->sampleRateComboBox->addItem("44100Hz", 44100);
 	ui->sampleRateComboBox->addItem("48000Hz", 48000);
 	ui->sampleRateComboBox->addItem("55466Hz", 55466);
+
+	struct Pair
+	{
+		SoundSource src;
+		int count;
+	};
+	static const QList<Pair> SRC_NUMS = {
+		{ SoundSource::FM, 6 }, { SoundSource::SSG, 3 }, { SoundSource::RHYTHM, 6 }, { SoundSource::ADPCM, 1 }
+	};
+
+	int tmpT = 0;
+	for (const auto& pair : SRC_NUMS) {
+		for (int i = 0; i < pair.count; ++i, ++tmpT) {
+			auto item = new QListWidgetItem(gui_utils::getTrackName(SongType::Standard, pair.src, i), ui->tracksListWidget);
+			item->setCheckState(std::any_of(defUnmutes.begin(), defUnmutes.end(), [tmpT](const int t) { return t == tmpT; })
+								? Qt::Checked : Qt::Unchecked);
+		}
+	}
 }
 
 WaveExportSettingsDialog::~WaveExportSettingsDialog()
@@ -52,4 +74,32 @@ int WaveExportSettingsDialog::getSampleRate() const
 int WaveExportSettingsDialog::getLoopCount() const
 {
 	return ui->loopSpinBox->value();
+}
+
+std::vector<int> WaveExportSettingsDialog::getSoloExportTracks() const
+{
+	if (!ui->tracksGroupBox->isChecked()) return {};
+
+	std::vector<int> tracks;
+	for (int i = 0; i < ui->tracksListWidget->count(); ++i) {
+		if (ui->tracksListWidget->item(i)->checkState() == Qt::Checked)
+			tracks.push_back(i);
+	}
+	return tracks;
+}
+
+void WaveExportSettingsDialog::on_reversePushButton_clicked()
+{
+	for (int i = 0; i < ui->tracksListWidget->count(); ++i) {
+		auto item = ui->tracksListWidget->item(i);
+		item->setCheckState((item->checkState() == Qt::Checked) ? Qt::Unchecked : Qt::Checked);
+	}
+}
+
+void WaveExportSettingsDialog::on_allPushButton_clicked()
+{
+	for (int i = 0; i < ui->tracksListWidget->count(); ++i) {
+		auto item = ui->tracksListWidget->item(i);
+		item->setCheckState(Qt::Checked);
+	}
 }